Remove static kernels from Model, leave them for tests and overpopulation
