+++ include/pops/kernel.hpp
@@ -46,112 +46,22 @@
 #ifndef POPS_KERNEL_HPP
 #define POPS_KERNEL_HPP
 
-#include "switch_kernel.hpp"
+#include "radial_kernel.hpp"
+#include "deterministic_kernel.hpp"
+#include "uniform_kernel.hpp"
+#include "neighbor_kernel.hpp"
+#include "network_kernel.hpp"
+#include "kernel_types.hpp"
+#include "kernel_base.hpp"
 #include "natural_anthropogenic_kernel.hpp"
 #include "config.hpp"
 
 namespace pops {
 
-/**
- * @brief Create natural kernel
- *
- * Kernel parameters are taken from the configuration.
- *
- * @param dispersers The disperser raster (reference, for deterministic kernel)
- * @param network Network (initialized or not)
- * @return Created kernel
- */
-template<typename IntegerRaster, typename RasterIndex>
-SwitchDispersalKernel<IntegerRaster, RasterIndex> create_static_natural_kernel(
-    const Config& config,
-    const IntegerRaster& dispersers,
-    const Network<RasterIndex>& network)
-{
-    auto natural_kernel = kernel_type_from_string(config.natural_kernel_type);
-    UniformDispersalKernel uniform_kernel(config.rows, config.cols);
-    RadialDispersalKernel<IntegerRaster> radial_kernel(
-        config.ew_res,
-        config.ns_res,
-        natural_kernel,
-        config.natural_scale,
-        direction_from_string(config.natural_direction),
-        config.natural_kappa,
-        config.shape);
-    DeterministicNeighborDispersalKernel natural_neighbor_kernel(
-        direction_from_string(config.natural_direction));
-    DeterministicDispersalKernel<IntegerRaster> deterministic_kernel(
-        natural_kernel,
-        dispersers,
-        config.dispersal_percentage,
-        config.ew_res,
-        config.ns_res,
-        config.natural_scale,
-        config.shape);
-    NetworkDispersalKernel<RasterIndex> network_kernel(
-        network, config.network_min_time, config.network_max_time);
-    SwitchDispersalKernel<IntegerRaster, RasterIndex> selectable_kernel(
-        natural_kernel,
-        radial_kernel,
-        deterministic_kernel,
-        uniform_kernel,
-        network_kernel,
-        natural_neighbor_kernel,
-        config.deterministic);
-    return selectable_kernel;
-}
-/**
- * @brief Create anthropogenic kernel
- *
- * Same structure as the natural kernel, but the parameters are for anthropogenic
- * kernel when available.
- *
- * @param dispersers The disperser raster (reference, for deterministic kernel)
- * @param network Network (initialized or not)
- * @return Created kernel
- */
-template<typename IntegerRaster, typename RasterIndex>
-SwitchDispersalKernel<IntegerRaster, RasterIndex> create_static_anthro_kernel(
-    const Config& config,
-    const IntegerRaster& dispersers,
-    const Network<RasterIndex>& network)
-{
-    auto anthro_kernel = kernel_type_from_string(config.anthro_kernel_type);
-    UniformDispersalKernel uniform_kernel(config.rows, config.cols);
-    RadialDispersalKernel<IntegerRaster> radial_kernel(
-        config.ew_res,
-        config.ns_res,
-        anthro_kernel,
-        config.anthro_scale,
-        direction_from_string(config.anthro_direction),
-        config.anthro_kappa,
-        config.shape);
-    DeterministicNeighborDispersalKernel anthro_neighbor_kernel(
-        direction_from_string(config.anthro_direction));
-    DeterministicDispersalKernel<IntegerRaster> deterministic_kernel(
-        anthro_kernel,
-        dispersers,
-        config.dispersal_percentage,
-        config.ew_res,
-        config.ns_res,
-        config.anthro_scale,
-        config.shape);
-    NetworkDispersalKernel<RasterIndex> network_kernel(
-        network, config.network_min_time, config.network_max_time);
-    SwitchDispersalKernel<IntegerRaster, RasterIndex> selectable_kernel(
-        anthro_kernel,
-        radial_kernel,
-        deterministic_kernel,
-        uniform_kernel,
-        network_kernel,
-        anthro_neighbor_kernel,
-        config.deterministic);
-    return selectable_kernel;
-}
-
 // TODO: Use smart pointers or ensure delete is called.
 
 /**
- * @brief Create natural kernel
+ * @brief Create natural kernel from configuration
  *
  * Kernel parameters are taken from the configuration.
  *
@@ -214,7 +124,7 @@ KernelInterface<Generator>* create_dynamic_natural_kernel(
 }
 
 /**
- * @brief Create anthropogenic kernel
+ * @brief Create anthropogenic kernel from configuration
  *
  * Same structure as the natural kernel, but the parameters are for anthropogenic
  * kernel when available.
@@ -300,28 +210,10 @@ KernelInterface<Generator>* create_dynamic_anthro_kernel(
  * See NaturalAnthropogenicDispersalKernel and SwitchDispersalKernel for further
  * documentation.
  */
-template<typename IntegerRaster, typename RasterIndex>
-using DispersalKernel = NaturalAnthropogenicDispersalKernel<
-    SwitchDispersalKernel<IntegerRaster, RasterIndex>,
-    SwitchDispersalKernel<IntegerRaster, RasterIndex>>;
-
 template<typename Generator>
 using DynamicDispersalKernel = DynamicNaturalAnthropogenicDispersalKernel<
     KernelInterface<Generator>,
     KernelInterface<Generator>>;
-
-template<typename IntegerRaster, typename RasterIndex>
-DispersalKernel<IntegerRaster, RasterIndex> create_static_kernel(
-    const Config& config,
-    const IntegerRaster& dispersers,
-    const Network<RasterIndex>& network)
-{
-    return DispersalKernel<IntegerRaster, RasterIndex>(
-        create_static_natural_kernel(config, dispersers, network),
-        create_static_anthro_kernel(config, dispersers, network),
-        config.use_anthropogenic_kernel,
-        config.percent_natural_dispersal);
-}
 
 template<typename Generator, typename IntegerRaster, typename RasterIndex>
 DynamicDispersalKernel<Generator> create_dynamic_kernel(
